refactor: share AES crypto helpers
